Moves `warnExactlyOneOf` to `stringvalidator.WarnExactlyOneOf`
