+++ internal/framework/validators/internal/exactly_one_of.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package s3
+package internal
 
 import (
 	"context"
@@ -15,15 +15,20 @@ import (
 	"github.com/hashicorp/terraform-provider-aws/internal/errs/fwdiag"
 )
 
-// This is a copy of `stringvalidator.ExactlyOneOf` and `schemavalidator.ExactlyOneOfValidator`
-// that returns a warning instead of an error.
-// It could likely be moved to an internal validators package if useful elsewhere.
-
-func warnExactlyOneOf(expressions ...path.Expression) validator.String {
-	return ExactlyOneOfValidator{
-		PathExpressions: expressions,
-	}
-}
+var (
+	// _ validator.Bool    = ExactlyOneOfValidator{}
+	// _ validator.Float32 = ExactlyOneOfValidator{}
+	// _ validator.Float64 = ExactlyOneOfValidator{}
+	// _ validator.Int32   = ExactlyOneOfValidator{}
+	// _ validator.Int64   = ExactlyOneOfValidator{}
+	// _ validator.List    = ExactlyOneOfValidator{}
+	// _ validator.Map     = ExactlyOneOfValidator{}
+	// _ validator.Number  = ExactlyOneOfValidator{}
+	// _ validator.Object  = ExactlyOneOfValidator{}
+	// _ validator.Set     = ExactlyOneOfValidator{}
+	_ validator.String = ExactlyOneOfValidator{}
+	// _ validator.Dynamic = ExactlyOneOfValidator{}
+)
 
 type ExactlyOneOfValidator struct {
 	PathExpressions path.Expressions
